Only have FFI code at the low level
